% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_supplant_new_data.R
\name{apply_to_new_data}
\alias{apply_to_new_data}
\title{Apply function to fitted object after supplanting data with new data}
\usage{
apply_to_new_data(
  FUN,
  fit,
  Xocc,
  Xobs = NULL,
  ModelSite = NULL,
  y = NULL,
  dataargs = NULL,
  funargs = NULL
)
}
\description{
Apply function to fitted object after supplanting data with new data
}
