% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fit.R
\name{artificial_runjags}
\alias{artificial_runjags}
\alias{simulate_covar_data}
\alias{simulate_iid_detections}
\title{Create a fully artificial fitted object}
\usage{
artificial_runjags(
  nspecies = 4,
  nsites = 100,
  nvisitspersite = 2,
  OccFmla = "~ UpSite + Sine1 + Sine2",
  ObsFmla = "~ UpVisit + Step",
  occ.b.min = -1,
  occ.b.max = 1,
  det.b.min = -1,
  det.b.max = 1,
  lv.b.min = -0.5,
  lv.b.max = 0.5,
  modeltype = "jsodm_lv",
  ...
)

simulate_covar_data(nsites, nvisitspersite)

simulate_iid_detections(nspecies, nvisits, p = 0.5)
}
\arguments{
\item{nspecies}{is the number of species to simulate. Species are named A, B, C... (max of 26 allowed)}

\item{nsites}{Number of ModelSites to simulate}

\item{nvisitspersite}{Number of visits per site (the same number for each site)}

\item{OccFmla}{Formula for occupancy. Available variables: UpSite, Sine1 and Sine2}

\item{ObsFmla}{Formula for detection. Available variables: Upvisit, Step}

\item{occ.b.min, }{occ.b.max, det.b.min, det.b.max The upper and lower bouonds of the occ.b and det.b parameters.
May be a single number or an array with rows corresponding to species and columns to covariates.}

\item{lv.b.min, }{lv.b.max Same as occ.b.min and occ.b.max for the latent variable loadings.}

\item{nvisits}{Number of visits in total to simulate}

\item{p}{The probability of detection, constant for all species and visits.}

\item{nlv}{Number of latent variables. Must be 4 or less}
}
\value{
A list that has enough similarities to runjags objects that residual calculations are possible.
The true parameter set is the first (and only row) of the first MCMC chain.
It can be accessed using get_theta(fit, type = 1)

A list with elements Xocc, and Xobs for the occupancy and detection covariates respectively

A simulated dataframe of detections. Column names are the species names, each row is a visit.
Elements of the data frame are either 1 (detected) or 0 (not detected).
}
\description{
Create a fully artificial fitted object
}
\details{
Occupancy covariate names are UpSite, Sine1 and Sine2.
Detection dovariate names are UpVisit and Step.

Every species is equally likely to be detected at every visit.
}
\section{Functions}{
\itemize{
\item \code{simulate_covar_data}: Generate fake covariate data.

\item \code{simulate_iid_detections}: Generate fake observation data.
}}

\examples{
artfit <- artificial_runjags(nspecies = 2, nsites = 10, nvisitspersite = 4, modeltype = "jsodm_lv", nlv = 2)
\# with high correlation between occupancy of species
artfit <- artificial_runjags(nspecies = 2, nsites = 10, nvisitspersite = 4,
                              OccFmla = "~ 1",
                              occ.b.min = 0.8,
                              lv.b.min = 0.3,
                              modeltype = "jsodm_lv",
                              nlv = 2)
 cor(artfit$data$y)
}
