% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edetperspecies.R
\name{Edetperspecies.jsodm}
\alias{Edetperspecies.jsodm}
\alias{Odetperspecies}
\title{Expected number of detections per species per site}
\usage{
Edetperspecies.jsodm(fit, desiredspecies = fit$species)

Odetperspecies(fit, y = NULL)
}
\arguments{
\item{fit}{A fitted jsodm}

\item{desiredspecies}{List of species names to include in the calculations}
}
\value{
An array of expected number of detections per modelsite x species x draw.
}
\description{
For each species individually calculates the expected number of detections at each ModelSite.
As this is the expected value, and each species treated individually, the same method can be applied to jsodm_lv models.
}
\section{Functions}{
\itemize{
\item \code{Odetperspecies}: Get observed detections per model site x species from the 'y' data in the fitted object.
}}

\examples{
fit <- readRDS("../sflddata/private/data/testdata/cutfit_7_4_11_2LV.rds")
fit <- translatefit(fit)
Endet <- Edetperspecies.jsodm(fit)
}
