% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{elpd}
\alias{elpd}
\title{Expected log predictive density: the probability density for each model site, average across posterior draws}
\usage{
elpd(likelmat)
}
\arguments{
\item{likelmat}{A matrix of predictive density for each draw (row) x modelsite (column)}
}
\description{
Expected log predictive density: the probability density for each model site, average across posterior draws
}
