% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcorrelations.R
\name{get.enviro.cor}
\alias{get.enviro.cor}
\title{Compute inter-species correlations due to similarity of the occupancy linear predictor at model sites}
\usage{
get.enviro.cor(fit, est = "median", prob = 0.95)
}
\arguments{
\item{fit}{An object created by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}}

\item{est}{Stastical summary of draws to return.}

\item{prob}{The target probability coverage of highest probability density intervals.}
}
\value{
A list of matrices. Column and row names are species names.
\code{cor} is the correlation estimate computed according to \code{est}
\code{cor.lower} is the lower bound of the highest posterior density interval for correlation
\code{cor.upper} is the upper bound of the highest posterior density interval for correlation
\code{sig.cor} is \code{cor} where all correlations with hpd that includes 0 are set to 0.
\code{cov} is the covariance estimate computed accordting to \code{est}
}
\description{
Compute inter-species correlations due to similarity of the occupancy linear predictor at model sites
}
\details{
For each draw from the posterior,
computes the correlation and covariance between the species-specific occupancy linear predictions without latent variables across all sites.
The \code{est}, for each species, is returned.
Significantly non-zero covariance or correlation is computed by \code{\link[coda:HPDinterval]{coda::HPDinterval()}} according to the signficance level \code{prob}.
}
\examples{
fit <- readRDS("../sflddata/private/data/testdata/cutfit_7_4_11_2LV.rds")
cor_envir <- get.enviro.cor(fit, est = "median", prob = 0.95)
}
