% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcorrelations.R
\name{get.residual.cor}
\alias{get.residual.cor}
\title{Compute inter-species correlations due to latent variable loadings}
\usage{
get.residual.cor(object, est = "median", prob = 0.95)
}
\arguments{
\item{est}{Stastical summary of draws to return.}

\item{prob}{The target probability coverage of highest probability density intervals.}

\item{fit}{An object created by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}}
}
\description{
Compute inter-species correlations due to latent variable loadings
}
\details{
This is modelled after the function \code{\link[boral:get.residual.cor]{boral::get.residual.cor()}}.
}
\examples{
fit <- readRDS("../sflddata/private/data/testdata/cutfit_7_4_11_2LV.rds")
cor_res <- get.residual.cor(fit, est = "median", prob = 0.95)
}
