% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_jsodm_lv_sepexp.R
\name{likelihood_LVvmarg_draw.jsodm_lv_sepexp}
\alias{likelihood_LVvmarg_draw.jsodm_lv_sepexp}
\title{Likelihood Computation for JSODM_LV Model for a Given Draw}
\usage{
likelihood_LVvmarg_draw.jsodm_lv_sepexp(
  Xocc,
  Xobs,
  y,
  ModelSite,
  occ.b,
  det.b,
  lv.b,
  lv.v.spatscale,
  lv.v.timescale,
  spatdistmat,
  timedistmat,
  zero.lv.v,
  numlvsims,
  simseed = NULL
)
}
\arguments{
\item{Xocc}{A matrix of processed occupancy covariate values for the model site. Must have 1 row.}

\item{Xobs}{A matrix of processed detection covariate values for each visit to the model site.}

\item{y}{Matrix of species detections for each visit to the model site.}

\item{occ.b_arr}{Occupancy covariate loadings. Each row is a species, each column an occupancy covariate, each layer (dim = 3) is a draw}

\item{det.b_arr}{Detection covariate loadings. Each row is a species, each column an detection covariate, each layer (dim = 3) is a draw}

\item{lv.b_arr}{LV loadings. Each row is a species, each column a LV, each layer (dim = 3) is a draw}

\item{data_i}{A row of a data frame created by \code{prep_data_by_modelsite}. Each row contains data for a single ModelSite.}

\item{lvsim}{A matrix of simulated LV values. Columns correspond to latent variables, each row is a simulation}
}
\description{
Computes the joint-species LV-marginal likelihood for all ModelSites, for a given draw.
}
