% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdetMackenzieBailey.R
\name{mb_stat}
\alias{mb_stat}
\title{Compare observed detection numbers to expected}
\usage{
mb_stat(fit, nperdraw = 2, seed = 122, simulations = NULL)
}
\description{
Expected detections compared to observed detections can be used to test model fit
(see MacKenzie and Bailey (2004) and Lemeshow, Sturdivant, Hosmer et al (2013)).
Note that the latter suggest that the p-value for the test is not very useful for comparing different models.
This function follows MacKenzie and Bailey (2004) by using cohorts given by
sites with the same number of survey visits. We further define cohorts by each species.
}
\details{
For each draw simulates LV twice and simulates detections for each of these once.
The expected number of detections is based on the assumpation that the full posterior is correct (and thus averages across posterior draws).
The distribution of test statistic according to the full posterior is obtained by simulation of the observations.
}
\examples{
fit <- readRDS("../sflddata/private/data/testdata/cutfit_7_4_11_2LV.rds")
fit <- translatefit(fit)
X2 <- mb_stat(fit)
}
