% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_fit_for_predict_only.R
\name{minimise_fit.jsodm_lv}
\alias{minimise_fit.jsodm_lv}
\title{Translate fit object into a smaller object for using in a webapp}
\usage{
minimise_fit.jsodm_lv(fit, ndrawsample = NULL)
}
\arguments{
\item{fit}{An object fitted by this package}

\item{ndrawsample}{If non-null then the number of posterior draw samples to keep, must be less than the number of posterior samples in \code{fit}.}
}
\value{
An object of the same class as \code{fit} with nearly all data removed. Only data used for predictions remains.
}
\description{
Translate fit object into a smaller object for using in a webapp
}
\examples{
fit <- readRDS("../Experiments/7_4_modelrefinement/fittedmodels/7_4_13_model_2lv_e13.rds")
fit <- msod:::translatefit(fit)
fit2 <- minimise_fit.jsodm_lv(fit)
saveRDS(fit2, file = "../farm_biodiversity_app/data/model_data.rds")
}
