% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesrichness_diff.R
\name{occspecrich_difference}
\alias{occspecrich_difference}
\title{difference in species richness to reference environmental variables.}
\usage{
occspecrich_difference(fit, Xocc, refXocc, sameLVvalues = TRUE)
}
\description{
difference in species richness to reference environmental variables.
}
\examples{
fit <- readRDS("../sflddata/private/data/testdata/cutfit_7_4_11_2LV.rds")
fit <- readRDS("../Experiments/7_4_modelrefinement/fittedmodels/7_4_13_model_2lv_e13.rds")
fit <- translatefit(fit)
Xocc <- unstandardise.designmatprocess(fit$XoccProcess, fit$data$Xocc[10, , drop = FALSE])
refXocc <- colMeans(unstandardise.designmatprocess(fit$XoccProcess, fit$data$Xocc[, , drop = FALSE]))
refXocc["NMdetected"] <- 1
refXocc <- t(refXocc)
refXocc <- as.data.frame(refXocc)
occspecrich_difference(fit, Xocc, refXocc)
refXocc <- Xocc
refXocc$woody500m <- 2
refXocc$NMdetected <- 1
refXocc$MaxTWarmMonth <- 370
occspecrich_difference(fit, Xocc, refXocc)
}
