% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterinitialisation.R
\name{paraminits}
\alias{paraminits}
\alias{paraminits.jsodm_lv}
\alias{paraminits.jsodm}
\title{Initial condition functions}
\usage{
paraminits(modeltype, chain, indata, ...)

paraminits.jsodm_lv(chain, indata, ...)

paraminits.jsodm(chain, indata, ...)
}
\arguments{
\item{chain}{Integer. Index of the chain.}

\item{indata}{A list of data that is typically passed to \code{\link[runjags:run.jags]{runjags::run.jags()}}}
}
\description{
Initial condition functions
}
\details{
Specifies the initial conditions for the MCMC chains.
}
\section{Functions}{
\itemize{
\item \code{paraminits.jsodm_lv}: Initialise paramers for the jsodm_lv model

\item \code{paraminits.jsodm}: Initialise parameters for the plain jsodm model
}}

