% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_param_names.R
\docType{data}
\name{parnames}
\alias{parnames}
\title{Convert old fit objects to use new parameter naming convention}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 23 rows and 2 columns.
}
\usage{
parnames
}
\description{
Change the variable names to new variables names
}
\examples{
model2lv <- readRDS("../Experiments/7_4_modelrefinement/fittedmodels/7_4_13_model_2lv_e13.rds")
model2lv_new <- translatefit(model2lv)
bestmodel_nolv <- readRDS("../Experiments/7_4_modelrefinement/fittedmodels/7_4_13_allhyp_vif_logwoody500m_msnm_year_Time_Wind.rds")
bestmodel_nolv_new <- translatefit(bestmodel_nolv)

}
\keyword{datasets}
