% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdetection_occupied.R
\name{pdet_occ_raw.jsodm}
\alias{pdet_occ_raw.jsodm}
\title{Probability of Detection, Given Occupied}
\usage{
pdet_occ_raw.jsodm(fixedcovar, loadfixed)
}
\arguments{
\item{fixedcovar}{An array of detection covariate values. Each row is a visit, each column is a covariate.}

\item{loadfixed}{An array of loadings for the covariates in 'fixedcovar'. Each row is a species,
each columns is a covariate (in same order as in fixedcovar), and each layer is a draw from the distribution of loadings.}
}
\value{
An array of detection probability values given species occupies the visited site. Each row is a visit, each column a species,
each layer a draw corresponding to the loadings.
}
\description{
Probability of Detection, Given Occupied
}
