% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdetection_occupied.R
\name{pdet_occ_raw.jsodm_lv_re}
\alias{pdet_occ_raw.jsodm_lv_re}
\title{Raw version for random effects in the detection takes a set of offsets
Random effect value is the value of a random effect for each row of the fixed covariates (in this case visits)
Can be constructed from a fitted object by \code{get_det_re(fit)[fit$data$SiteObsInd, ]}}
\usage{
pdet_occ_raw.jsodm_lv_re(fixedcovar, loadfixed, det.re)
}
\description{
Raw version for random effects in the detection takes a set of offsets
Random effect value is the value of a random effect for each row of the fixed covariates (in this case visits)
Can be constructed from a fitted object by \code{get_det_re(fit)[fit$data$SiteObsInd, ]}
}
