% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poccupancy.R
\name{poccupy.jsodm_lv}
\alias{poccupy.jsodm_lv}
\title{lvvfromposterior = TRUE draws lv.v from the posterior with all other loadings.}
\usage{
\method{poccupy}{jsodm_lv}(fit, usethetasummary = NULL, lvvfromposterior = TRUE, margLV = FALSE)
}
\description{
lvvfromposterior = TRUE draws lv.v from the posterior with all other loadings.
}
