% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJAGSdata.R
\name{prepJAGSdata}
\alias{prepJAGSdata}
\alias{prepJAGSdata.jsodm}
\alias{prepJAGSdata.jsodm_lv}
\alias{prepJAGSdata.jsodm_lv_sepexp}
\alias{prep_new_data}
\title{Prepare JAGS-like Data for JSODM}
\usage{
prepJAGSdata(
  modeltype,
  Xocc,
  yXobs,
  ModelSite,
  species,
  XoccProcess,
  XobsProcess,
  ...
)

prepJAGSdata.jsodm(
  Xocc,
  yXobs,
  ModelSite,
  species,
  XoccProcess,
  XobsProcess,
  ...
)

prepJAGSdata.jsodm_lv(
  Xocc,
  yXobs,
  ModelSite,
  species,
  XoccProcess,
  XobsProcess,
  nlv
)

prepJAGSdata.jsodm_lv_sepexp(
  Xocc,
  yXobs,
  ModelSite,
  species,
  XoccProcess,
  XobsProcess,
  nlv,
  SpatDist,
  TimeDist,
  ...
)

prep_new_data(fit, Xocc, yXobs, ModelSite, ...)
}
\arguments{
\item{Xocc}{A dataframe of covariates related to occupancy. One row per ModelSite.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{yXobs}{A dataframe of species observations (1 or 0) and covariates related to observations. One row per visit.
Each column is either a covariate or a species.
Must also include the ModelSiteVars columns to uniquely specify ModelSite.}

\item{ModelSite}{A list of column names in y, Xocc and Xobs that uniquely specify the ModelSite. Can be simply a ModelSite index}

\item{species}{A list of species names (corresponding to columns of yXobs) to model.}

\item{XoccProcess}{An object create by prep.designmatprocess for the occupancy covariates}

\item{XobsProcess}{An object create by prep.designmatprocess for the observation covariates}

\item{nlv}{The number of latent variables}

\item{OccFmla}{A formula specifying the occupancy model in terms of the covariates in Xocc.}

\item{ObsFmla}{A formula specifying the model for detection in terms of the covariates in Xobs.}
}
\description{
Prepare JAGS-like Data for JSODM
}
\details{
Given the input data parameters of run.detectionoccupancy prepare the data list for JAGS
}
\section{Functions}{
\itemize{
\item \code{prepJAGSdata.jsodm}: Data preparation for the jsodm model.

\item \code{prepJAGSdata.jsodm_lv}: Data preparation for the jsodm_lv model

\item \code{prepJAGSdata.jsodm_lv_sepexp}: Data preparation for the jsodm_lv_sepexp model (latent variables with separable covariance that is an exponential function)

\item \code{prep_new_data}: A short function that applies the prepJAGSdata function to new data, given an object created by run.detectionoccupancy
Xocc, yXobs, ModelSite must follow some rules as for run.detectionoccupancy, except yXobs may omit the species detections
}}

