% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJAGSdata_nostandardising.R
\name{prepJAGSdata2}
\alias{prepJAGSdata2}
\title{Prepare JAGS-like Data for JSODM}
\usage{
prepJAGSdata2(modeltype, Xocc, Xobs, y, ModelSite, ...)
}
\arguments{
\item{Xocc}{A dataframe of covariates related to occupancy. One row per ModelSite.}

\item{Xobs}{A dataframe of covariates related to observations. Each column is a covariate.}

\item{y}{A dataframe of species observations (1 or 0). One row per visit. Each column is a species.}

\item{ModelSite}{A list of integers giving the number each row of Xocc (ModelSite) visited for each visit (row of y and Xobs)}

\item{nlv}{The number of latent variables}
}
\description{
Prepare JAGS-like Data for JSODM
}
\details{
Given the input data parameters of run.detectionoccupancy prepare the data list for JAGS
}
