% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmanybern.R
\name{rmanybern}
\alias{rmanybern}
\title{Simulate many independent Bernoulli random variables}
\usage{
rmanybern(probarr)
}
\arguments{
\item{probarr}{An array of probability of success.
Each element of the array represents the probability of success of a Bernoulli random variable that is independent of all other random variables.}
}
\value{
An array the same dimensions of probarr with simulated values of TRUE (success) or FALSE (failure).
}
\description{
Simulate many independent Bernoulli random variables
}
