% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_fit.R
\name{simulate_detections}
\alias{simulate_detections}
\alias{simulate_detections_lv.v}
\title{Simulate observations from parameters of a fitted object.}
\usage{
simulate_detections(fit, esttype = "median", seed = NULL)

simulate_detections_lv.v(fit, esttype = "median", seed = NULL)
}
\arguments{
\item{fit}{A runjags fitted object created by \code{\link[=run.detectionoccupancy]{run.detectionoccupancy()}}}

\item{esttype}{Specifies parameter set to extract from fit. See \code{\link[=get_occ_b]{get_occ_b()}} and similar for specification.}
}
\description{
Simulate observations from parameters of a fitted object.
}
\details{
Simulate detections from all given parameters (including LV and random effects)
}
\section{Functions}{
\itemize{
\item \code{simulate_detections_lv.v}: Simulate LV values and detections.
Uses \code{\link[=simulate_detections]{simulate_detections()}} and \code{\link[=simulate_lv.v]{simulate_lv.v()}}.
For each site the simulated LV values are copied across all draws.
}}

\examples{
fit <- readRDS("./tmpdata/deto_wind.rds")
fit$data <- as_list_format(fit$data)
class(fit) <- c("jsodm", class(fit))
detected <- simulate_detections(fit, esttype = "median", UseFittedLV = TRUE)

}
