% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesrichness_morelikelythannot.R
\name{speciesrichness_mltn}
\alias{speciesrichness_mltn}
\title{Species richness as number of species more likely than not to occur on the patch / farm.}
\usage{
speciesrichness_mltn(fit, Xocc)
}
\description{
For each draw of lvv and parameters, computes the number of species with any probability in any patch is greater than 0.5
}
\examples{
fit <- readRDS("../Experiments/7_4_modelrefinement/fittedmodels/7_4_13_model_2lv_e13.rds")
fit <- translatefit(fit)
Xocc <- unstandardise.designmatprocess(fit$XoccProcess, fit$data$Xocc[c(1, 500), , drop = FALSE])
speciesrichness_mltn(fit, Xocc)
}
