% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesrichness.R
\name{prednumsuccess}
\alias{prednumsuccess}
\title{Expected Biodiversity}
\usage{
prednumsuccess(probarr)
}
\description{
The expected number of species occupying a ModelSite, or the expected number of species detected at a ModelSite.
}
